requirements.txt
accelerate==1.10.0
diffusers==0.35.0
einops==0.8.1
imageio==2.37.0
numpy==2.3.4
omegaconf==2.3.0
Pillow==12.0.0
pyspng==0.1.3
Requests==2.32.5
safetensors==0.6.2
scipy==1.16.3
tensorflow==2.16.1
timm==1.0.21
tqdm==4.67.1
wandb==0.22.3
